requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
face_recognition==1.3.0
pytest==8.3.3
python-multipart==0.0.9
